backup_20251026_033054/requirements.txt
streamlit==1.32.0
edge-tts==6.1.18
faster-whisper==1.0.3
Pillow==9.5.0
numpy==1.26.0
pydub==0.25.1
scipy==1.11.4